requirements.txt
discord.py
httpx
beautifulsoup4
python-dotenv
apscheduler
